Fix MAX_ROTTE_AR definition to ensure proper type casting

+++ Codice.c
@@ -37,7 +37,7 @@ Consegna:   Movhex è una compagnia di autotrasporti che dispone di una flotta d
 #define FALSO "KO"
 #define NOT_VALID_COST 0
 #define LUNGHEZZA_STR_COMANDO_MAX 16
-#define MAX_ROTTE_AR 5
+#define MAX_ROTTE_AR (u_int16_t)5
 #define NOT_VALID (u_int16_t)-2
 #define NON_USCENTE 1       //per definire se un esagono è destinazione o partenza di una rotta aerea 
 #define GRIGIO 7
